+++ SolastaUnfinishedBusiness/Translations/en/SubClasses/CollegeOfHarlequin-en.txt
@@ -1,6 +1,6 @@
 ﻿Feedback/&BardicInspirationUsedToBoostCombatAbility={0} spends a {1} inspiration die to increase attack bonus, attack damage, AC and movement speed by {2}
 Feature/&CombatInspirationTitle=Combat Inspiration
-Feature/&CombatInspirationDescription=You are inspired when bathing in the heat of a fight. You can roll a bardic inspiration die with advantage and add the roll to your weapon attack, damage, AC and movement speed. The inspiration lasts until the start of your next turn
+Feature/&CombatInspirationDescription=You are inspired when bathing in the heat of a fight. You can roll a bardic inspiration die with advantage and add the roll to your weapon attack, damage, AC and movement speed. The inspiration lasts until the start of your next turn.
 Condition/&ConditionFightingAbilityEnhancedTitle=Inspired
 Condition/&ConditionFightingAbilityEnhancedDescription=Your fighting ability is greatly enhanced.
 Feature/&CombatInspirationArmorClassEnhancementTitle=Inspired Defense
@@ -13,5 +13,7 @@ Feature/&TerrificPerformanceTitle=Terrific Performance
 Feature/&TerrificPerformanceDescription=You strike down your enemies in a magnificent way that inspired fear in those who are still alive. Whenever you kill an enemy all other enemies within 15' of you must make a WIS saving throw against your spell DC to avoid being terrified. A terrified creature has disadvantage on attack and saving throw. The effect last until the end of your next turn.
 Feature/&ImproveCombatInspirationTitle=Improved Combat Inspiration
 Feature/&ImproveCombatInspirationDescription=You regain a bardic inspiration die whenever you strike down an enemy.
+Feature/&ProficiencyCollegeOfHarlequinMartialWeaponTitle=Martial Art Training
+Feature/&ProficiencyCollegeOfHarlequinMartialWeaponDescription=Proficient with all simple and martial art weapons.
 Subclass/&CollegeOfHarlequinTitle=College of Harlequin
 Subclass/&CollegeOfHarlequinDescription=The Harlequins see no distinction between art and war. Bards of this college see it is their duty to keep the legend of the great battle of the Cataclysm alive through their terrific performance, either on stage or on the battle field.
